rename plugins (remove 'inject' prefix)
