+++ types/shims-vue.d.ts
@@ -6,7 +6,7 @@ import {
 } from 'vuex';
 import dayjs from 'dayjs';
 import { NuxtAxiosInstance } from '@nuxtjs/axios';
-import { Toast } from '~/plugins/inject-toast';
+import { Toast } from '~/plugins/toast';
 
 import { SpotifyEndpoints } from '~/plugins/spotify/endpoints';
